extract SidebarShell from the old sidebar
